backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]
sqlmodel
psycopg[binary]
pydantic-settings
python-jose[cryptography]
httpx
alembic
